requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
Mako==1.3.5
MarkupSafe==3.0.1
openai==1.51.2
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.4
